+++ 05_command/DocumentEditor/InsertParagraphCommand.h
@@ -8,7 +8,7 @@ class CInsertParagraphCommand : public CAbstractCommand
 {
 public:
 	CInsertParagraphCommand(std::vector<CDocumentItem> & items, 
-		std::shared_ptr<IParagraph> paragraph, boost::optional<unsigned> position);
+		std::shared_ptr<IParagraph> paragraph, boost::optional<size_t> position);
 protected:
 	void DoExecute() override;
 	void DoUnexecute() override;
@@ -16,6 +16,6 @@ protected:
 private:
 	std::vector<CDocumentItem> & m_items;
 	CDocumentItem m_item;
-	boost::optional<unsigned> m_position;
+	boost::optional<size_t> m_position;
 };
